Citation Explorer: extract item tree into its own file
Move the shared row classes, row provider, and ItemTree subclass out
of citationExplorer.js without changing their behavior.
